old-tests/submission/W3C/domhtml/HTMLIFrameElement05.html

<!DOCTYPE html>
<html><head>
<meta charset="utf-8">
<title>NIST DOM HTML Test - IFRAME</title>
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<link href="http://www.w3.org/TR/html5/obsolete.html#dom-iframe-marginwidth" rel="help">
<link rel="author" title="NIST">
</head>
<body>
<iframe height="50" marginwidth="5" src="/support/html/sample.html" width="60">IFRAME1</iframe>


<script>
var domtest = async_test("The marginWidth IDL attribute of the frame element must reflect the element's marginwidth content attribute.");

on_event(window, "load", domtest.step_func(function (dom_event) {
    var nodeList;
      var testNode;
      var vmarginwidth;
      var doc;
      doc = window.document;
      nodeList = doc.getElementsByTagName("iframe");
      assert_equals(nodeList.length,1);
testNode = nodeList.item(0);
      vmarginwidth = testNode.marginWidth;

      assert_equals(vmarginwidth,"5");
       
    domtest.done();
}));

</script>
<p>The marginWidth IDL attribute of the frame element must reflect the element's marginwidth content attribute.</p>
<div id="log"></div>
</body></html>
